+++ packages/search/projects/search-lib/src/public-api.ts
@@ -1,0 +1,8 @@
+/*
+ * Public API Surface of my-lib
+ */
+
+export * from './lib/search-lib.module';
+export * from './lib/search/search.component';
+export * from './lib/lib-configuration';
+export * from './lib/types';