+++ sandbox/workflow-ms-example/CHANGELOG.md
@@ -3,6 +3,16 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## 0.3.0 (2021-09-12)
+
+* feat(sandbox): basic-ci-setup (#298) ([fc50e81](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/fc50e81)), closes [#298](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/298)
+* feat(sandbox): example of pubnub as a provider in notification service (#279) ([e30a0ed](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/e30a0ed)), closes [#279](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/279) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#0](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/0) [#278](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/278) [#211](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/211) [#263](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/263) [#280](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/280) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#298](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/298) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#263](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/263) [#280](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/280) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208) [#208](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/208)
+* chore(deps): bump path-parse in /sandbox/workflow-ms-example (#305) ([5aefab1](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/5aefab1)), closes [#305](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/305)
+
+
+
+
+
 ## 0.2.0 (2021-08-02)
 
 * feat(sandbox): k8s-yaml-to-hcl (#280) ([621f0cc](https://github.com/sourcefuse/loopback4-microservice-catalog/commit/621f0cc)), closes [#280](https://github.com/sourcefuse/loopback4-microservice-catalog/issues/280)